requirements.txt
django~=4.2.11
Pillow~=10.3.0
django-mptt~=0.16.0
django-colorfield~=0.11.0
